refactor(consensus): read validators from genesis file
